requirements.txt
pymongo
python-telegram-bot
requests
beautifulsoup4
